+++ docs/CHANGELOG.rst
@@ -18,6 +18,7 @@ API Changes
   - **Change:** Reprojection has improved performance due to one less shuffle stage and lower memory usage.
   ``TileRDDReproject`` loses dependency on ``TileReprojectMethods`` in favor of ``RasterRegionReproject``
   - **New:** CollectionLayerReader now has an SPI interface.
+  - **New:** ``ZoomResample`` can now be used on ``MultibandTileLayerRDD``\s.
 
 1.2.1
 _____